requirements.txt
# 硕方 T50Pro 标签打印机控制程序依赖

# 串口通信
pyserial>=3.5

# 蓝牙 BLE
bleak>=0.20.0

# 图像处理
Pillow>=9.0.0

# 条形码生成
python-barcode>=0.15.0

# 二维码生成
qrcode[pil]>=7.0

# 开发依赖（测试用）
pytest>=7.0.0
pytest-asyncio>=0.21.0
